Sistema biblioteca finalizado

+++ routes/estudiantes.py
@@ -11,13 +11,27 @@ def estudiantes():
 
     cursor = conexion.cursor()
 
-    cursor.execute("SELECT * FROM estudiantes")
+    buscar = request.args.get('buscar', '')
+
+    error = request.args.get('error')
+
+    cursor.execute("""
+        SELECT *
+        FROM estudiantes
+        WHERE nombre LIKE ?
+        OR cedula LIKE ?
+    """, (
+        '%' + buscar + '%',
+        '%' + buscar + '%'
+    ))
 
     estudiantes = cursor.fetchall()
 
     return render_template(
         'estudiantes.html',
-        estudiantes=estudiantes
+        estudiantes=estudiantes,
+        buscar=buscar,
+        error=error
     )
 
 
@@ -26,6 +40,9 @@ def agregar_estudiante():
 
     nombre = request.form['nombre']
     cedula = request.form['cedula']
+    if len(cedula) < 9:
+
+        return redirect('/estudiantes')
     correo = request.form['correo']
     carrera = request.form['carrera']
     
@@ -48,12 +65,21 @@ def eliminar_estudiante(id):
 
     cursor = conexion.cursor()
 
+    # VALIDAR PRESTAMOS
+    cursor.execute("""
+        SELECT COUNT(*) AS total
+        FROM prestamos
+        WHERE id_estudiante = ?
+    """, (id))
+    prestamos = cursor.fetchone()
+    if prestamos.total > 0:
+        return redirect('/estudiantes?error=prestamos')
+
+    # ELIMINAR
     cursor.execute("""
         DELETE FROM estudiantes
         WHERE id_estudiante = ?
     """, (id))
-
     conexion.commit()
-
     return redirect('/estudiantes')
